requirements-dev.txt
pytest>=8.0.0
pytest-cov>=4.1.0
ruff>=0.3.0
mypy>=1.8.0
pandas-stubs>=2.1.4
types-PyYAML>=6.0.0
